+++ Lab.Net.EF/Lab.Net.EF.MVC/Controllers/EmpleadosController.cs
@@ -83,7 +83,7 @@ namespace Lab.Net.EF.MVC.Controllers
             return View("InsertarModificar");
         }
 
-        public ActionResult Eliminar(decimal id)
+        public ActionResult Eliminar(int id)
         {
             try
             {
