+++ Lab.Net.EF/Lab.Net.EF.MVC/Controllers/ProveedoresController.cs
@@ -84,7 +84,7 @@ namespace Lab.Net.EF.MVC.Controllers
             return View("InsertarModificar");
         }
 
-        public ActionResult Eliminar(decimal id)
+        public ActionResult Eliminar(int id)
         {
             try
             {
